Fix a bug in method numBytesToEncode in PimIndexWriter
